Restructure feature polling_stations
