--- frontend/src/api/polling_station/index.ts
@@ -1,2 +1,0 @@
-export * from "./usePollingStationListRequest";
-export * from "./usePollingStationGet";
