+++ frontend/src/features/polling_stations/components/PollingStationsLayout.tsx
@@ -1,8 +1,7 @@
 import { Outlet } from "react-router";
 
 import { Footer } from "@/components/footer/Footer";
-
-import { AppLayout } from "@kiesraad/ui";
+import { AppLayout } from "@/components/ui";
 
 export function PollingStationsLayout() {
   return (
