+++ frontend/src/features/polling_stations/hooks/usePollingStationListRequest.ts
@@ -1,8 +1,4 @@
-import {
-  POLLING_STATION_LIST_REQUEST_PATH,
-  PollingStationListResponse,
-  useInitialApiGetWithErrors,
-} from "@kiesraad/api";
+import { POLLING_STATION_LIST_REQUEST_PATH, PollingStationListResponse, useInitialApiGetWithErrors } from "@/api";
 
 export function usePollingStationListRequest(electionId: number) {
   const path: POLLING_STATION_LIST_REQUEST_PATH = `/api/elections/${electionId}/polling_stations`;
