+++ frontend/src/lib/util/hook/index.ts
@@ -1,5 +1,4 @@
 export * from "./useDebouncedCallback";
-export * from "./useForm";
 export * from "./useNumericParam";
 export * from "./usePollingStationStatus";
 export * from "./usePreventFormEnterSubmit";
